Load the tokenizer from a protobuf file
Like Alpino, this crate used to compile the tokenizer to a large C
object file. While this simplified use of the tokenizer, it has
several large downsides:

- The compilation time of the alpino-tokenizer-sys crate was long.
- Applications that use alpino-tokenizer end up with bloated binaries.
- The tokenizer uses a lot of memory, even if we do not use it (e.g.
  in multi-lingual applications).
- Since the tokenizer talks to C-land we need to use `unsafe`, but
  not in a way that is guaranteed to be safe.

With this change, we instead load the tokenizer from a simple protobuf
file that stores the transducer. It resolves all of the problems above
and allows us to perform tokenization in pure Rust.